# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Call-a-Vet Gateway API
  version: 0.1.0
  description: >
    Public+private HTTP API for Call-a-Vet. Auth via Supabase JWT (Bearer).
    Primary login is Phone OTP (Supabase Auth: send code → verify code → session).
    RLS enforced at the DB layer. Idempotency on critical mutations via `Idempotency-Key`.
servers:
  - url: https://api.dev.callavet.mx
    description: Development
  - url: https://api.staging.callavet.mx
    description: Staging
  - url: https://api.callavet.mx
    description: Production
...
paths:
...
  /pets:
    get:
      tags: [Pets]
      summary: List my pets
      security: [{ bearerAuth: [] }]
      responses: { "200": { $ref: "#/components/responses/ListPets" } }
    post:
      tags: [Pets]
      summary: Create a pet
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/PetUpsert" }
      responses: { "201": { $ref: "#/components/responses/Pet" } }
  /pets/{petId}:
    get:
      tags: [Pets]
      summary: Get a pet
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: "#/components/parameters/petId" } ]
      responses: { "200": { $ref: "#/components/responses/Pet" } }
    patch:
      tags: [Pets]
      summary: Update a pet
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: "#/components/parameters/petId" } ]
      requestBody: { $ref: "#/components/requestBodies/PetUpsert" }
      responses: { "200": { $ref: "#/components/responses/Pet" } }
    delete:
      tags: [Pets]
      summary: Archive a pet
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: "#/components/parameters/petId" } ]
      responses: { "204": { description: No Content } }
  /pets/{petId}/files/signed-url:
    post:
      tags: [Files, Pets]
      summary: Create signed upload URL for pet assets
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: "#/components/parameters/petId" } ]
      requestBody: { $ref: "#/components/requestBodies/SignedUrl" }
      responses: { "200": { $ref: "#/components/responses/SignedUrl" } }
...
components:
  securitySchemes:
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    petId: { in: path, name: petId, required: true, schema: { type: string, format: uuid } }
...
  requestBodies:
...
    PetUpsert:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PetInput"
    SignedUrl:
      content:
        application/json:
          schema:
            type: object
            properties:
              path: { type: string }
              contentType: { type: string }
...
  responses:
...
    ListPets:
      description: Pets
      content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/Pet" } } } } } }
    Pet:
      description: Pet
      content: { application/json: { schema: { $ref: "#/components/schemas/Pet" } } }
...
    SignedUrl:
      description: Signed URL response
      content:
        application/json:
          schema:
            type: object
            properties:
              path: { type: string }
              url: { type: string, format: uri }
              token: { type: string, nullable: true }
              expires_in: { type: integer, nullable: true }
              expiresIn: { type: integer, nullable: true }
              method: { type: string }
              headers: { type: object, additionalProperties: true }
...
  schemas:
...
    PetInput:
      type: object
      properties:
        name: { type: string, maxLength: 100 }
        species: { type: string, enum: [horse] }
        sex: { type: string, enum: [male, female, gelding] }
        age_range: { type: string, enum: [foal_0_2, young_3_5, adult_6_15, senior_16_plus] }
        weight_range: { type: string, enum: [lt_400, 400_500, 500_600, gt_600] }
        location:
          type: object
          properties:
            country: { type: string, maxLength: 100 }
            state_region: { type: string, maxLength: 100 }
        location_country: { type: string, maxLength: 100 }
        location_state_region: { type: string, maxLength: 100 }
        breed: { type: string, enum: [quarter_horse, thoroughbred, pre, arabian, criollo, appaloosa, paint_horse, warmblood, mixed, other] }
        warmblood_subbreed: { type: string, enum: [holsteiner, hanoverian, kwpn, oldenburg, selle_francais, westphalian, trakehner, other], nullable: true }
        other_breed_text: { type: string, maxLength: 100, nullable: true }
        primary_activity: { type: string, enum: [competition, regular_training, rehabilitation_recovery, retired, recreational] }
        discipline: { type: string, enum: [jumping, dressage, polo, endurance, barrel_racing, reining, charreada, ranch_work, recreational, other] }
        other_discipline_text: { type: string, maxLength: 100, nullable: true }
        training_intensity: { type: string, enum: [1_2_per_week, 3_4_per_week, 5_plus_per_week] }
        terrain: { type: string, enum: [sand, grass, dirt, mixed, other] }
        other_terrain_text: { type: string, maxLength: 100, nullable: true }
        observed_last_6_months:
          type: array
          items: { type: string, enum: [mild_lameness, stiffness, performance_drop, appetite_changes, none] }
        known_conditions:
          type: array
          items: { type: string, enum: [digestive, locomotor, respiratory, skin, none] }
        current_treatments_or_supplements: { type: string, maxLength: 500, nullable: true }
        last_vet_check: { type: string, enum: [lt_3_months, 3_6_months, gt_6_months, dont_remember] }
        vaccines_up_to_date: { type: string, enum: [yes, no, not_sure] }
        deworming_status: { type: string, enum: [regular, irregular, not_sure] }
        additional_notes: { type: string, maxLength: 1000, nullable: true }
    Pet:
      allOf:
        - $ref: "#/components/schemas/PetInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            user_id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
...